Create authController with functions login and logount
Afte proper authentication, we save the logged-in users data in the session: req.session.loggedUser = emp;

This way, on subsequent requests from this user, we will be able to recognize him/her and, for example, display his/hers data on the page, and protect our application from unauthorized use.

The logout function removes the users data from the session.