Adding a function that makes session data available to templates
The session data is not directly available in the page templates, so we will write a function to make it available.

The res.locals is the set of all parameters passed to the template. The EJS template engine will throw an error, if we use any parameter that was not previously passed. The above function protects us from this error. In addition to the session data, we also need to protect the parameter in which the login error is passed.